Move all commands into submodule
