+++ mkdocs/tests/build_tests.py
@@ -15,7 +15,8 @@ except ImportError:
     pass
 
 
-from mkdocs import build, nav, config
+from mkdocs import nav, config
+from mkdocs.commands import build
 from mkdocs.exceptions import MarkdownNotFound
 from mkdocs.tests.base import dedent
