+++ mkdocs/tests/cli_tests.py
@@ -15,7 +15,7 @@ class CLITests(unittest.TestCase):
     def setUp(self):
         self.runner = CliRunner()
 
-    @mock.patch('mkdocs.serve.serve', autospec=True)
+    @mock.patch('mkdocs.commands.serve.serve', autospec=True)
     def test_serve(self, mock_serve):
 
         result = self.runner.invoke(
@@ -24,7 +24,7 @@ class CLITests(unittest.TestCase):
         self.assertEqual(result.exit_code, 0)
         self.assertEqual(mock_serve.call_count, 1)
 
-    @mock.patch('mkdocs.build.build', autospec=True)
+    @mock.patch('mkdocs.commands.build.build', autospec=True)
     def test_build(self, mock_build):
 
         result = self.runner.invoke(
@@ -33,7 +33,7 @@ class CLITests(unittest.TestCase):
         self.assertEqual(result.exit_code, 0)
         self.assertEqual(mock_build.call_count, 1)
 
-    @mock.patch('mkdocs.build.build', autospec=True)
+    @mock.patch('mkdocs.commands.build.build', autospec=True)
     def test_build_verbose(self, mock_build):
 
         result = self.runner.invoke(
@@ -42,7 +42,7 @@ class CLITests(unittest.TestCase):
         self.assertEqual(result.exit_code, 0)
         self.assertEqual(mock_build.call_count, 1)
 
-    @mock.patch('mkdocs.build.build', autospec=True)
+    @mock.patch('mkdocs.commands.build.build', autospec=True)
     def test_json(self, mock_build):
 
         result = self.runner.invoke(
@@ -51,7 +51,7 @@ class CLITests(unittest.TestCase):
         self.assertEqual(result.exit_code, 0)
         self.assertEqual(mock_build.call_count, 1)
 
-    @mock.patch('mkdocs.new.new', autospec=True)
+    @mock.patch('mkdocs.commands.new.new', autospec=True)
     def test_new(self, mock_new):
 
         result = self.runner.invoke(
@@ -60,7 +60,7 @@ class CLITests(unittest.TestCase):
         self.assertEqual(result.exit_code, 0)
         self.assertEqual(mock_new.call_count, 1)
 
-    @mock.patch('mkdocs.gh_deploy.gh_deploy', autospec=True)
+    @mock.patch('mkdocs.commands.gh_deploy.gh_deploy', autospec=True)
     def test_gh_deploy(self, mock_gh_deploy):
 
         result = self.runner.invoke(
